+++ src/main/java/com/mailgun/client/MailgunClient.java
@@ -4,6 +4,7 @@ import java.util.concurrent.Executors;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mailgun.api.MailgunApi;
+import com.mailgun.form.FormEncoder;
 import com.mailgun.util.ConsoleLogger;
 import com.mailgun.util.MailgunApiUtil;
 import com.mailgun.util.ObjectMapperUtil;
@@ -16,7 +17,6 @@ import feign.Request;
 import feign.Retryer;
 import feign.auth.BasicAuthRequestInterceptor;
 import feign.codec.ErrorDecoder;
-import feign.form.FormEncoder;
 import feign.jackson.JacksonDecoder;
 import feign.jackson.JacksonEncoder;
 import feign.querymap.FieldQueryMapEncoder;