+++ CHANGELOG.md
@@ -5,14 +5,18 @@ All _notable_ changes to this project will be documented in this file.
 The format is based on _[Keep a Changelog][keepachangelog]_, and this project
 adheres to _[Semantic Versioning][semver]_.
 
+## [1.0.5] (released: 2022-11-09)
+### Added
+- multiple FormData attachment support.
+
 ## [1.0.4] (released: 2022-10-10)
 ### Added
-- add FormData attachment support.
-- add default AsyncClient.
+- FormData attachment support.
+- default AsyncClient.
 
 ## [1.0.3] (released: 2022-09-23)
 ### Added
-- add attachments support.
+- attachments support.
 
 ## [1.0.2] (released: 2022-08-28)
 ### Added
@@ -38,6 +42,7 @@ adheres to _[Semantic Versioning][semver]_.
 - Add Import a list of bounces from CSV file API
 
 
+[1.0.5]: https://github.com/mailgun/mailgun-java/compare/release/1.0.4...release/1.0.5
 [1.0.4]: https://github.com/mailgun/mailgun-java/compare/release/1.0.3...release/1.0.4
 [1.0.3]: https://github.com/mailgun/mailgun-java/compare/release/1.0.2...release/1.0.3
 [1.0.2]: https://github.com/mailgun/mailgun-java/compare/release/1.0.1...release/1.0.2